requirements.txt
Django>=1.7
django-roughpages
django-angular
arrow
django-braces
djangorestframework
